requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
singlestoredb>=1.0.0
